Split LogView into subviews
